+++ Filmoon.WebAPI/Services/ScreenwritersService.cs
@@ -29,42 +29,48 @@ public class ScreenwritersService
         {
             await ScreenwritersRepository.AddAsync(screenwriter);
 
-            return new ActionResponse(true, "Screenwriter was added successfully.");
+            return new ActionResponse(true, "Screenwriter was added successful.");
         }
 
         catch (Exception exception)
         {
-            return new ActionResponse(false, $"Screenwriter was not added. An exception occured: {exception.Message ?? exception.InnerException?.Message ?? "Undefined message"}.");
+            return new ActionResponse(false, $"Screenwriter was not added successful. An exception occured: {exception.Message ?? exception.InnerException?.Message ?? "Undefined message"}.");
         }
     }
 
     public async Task<ActionResponse> UpdateAsync(ScreenwriterEntity screenwriter)
     {
+        if (await ScreenwritersRepository.GetByIdAsync(screenwriter.Id) is null) return new ActionResponse(false, $"Screenwriter with Id {screenwriter.Id} was not found.");
+
         try
         {
             await ScreenwritersRepository.UpdateAsync(screenwriter);
 
-            return new ActionResponse(true, "Screenwriter was updated successfully.");
+            return new ActionResponse(true, "Screenwriter was updated successful.");
         }
 
         catch (Exception exception)
         {
-            return new ActionResponse(false, $"Screenwriter was not updated. An exception occured: {exception.Message ?? exception.InnerException?.Message ?? "Undefined message"}.");
+            return new ActionResponse(false, $"Screenwriter was not updated successful. An exception occured: {exception.Message ?? exception.InnerException?.Message ?? "Undefined message"}.");
         }
     }
 
-    public async Task<ActionResponse> RemoveAsync(ScreenwriterEntity screenwriter)
+    public async Task<ActionResponse> RemoveAsync(int id)
     {
+        var screenwriter = await ScreenwritersRepository.GetByIdAsync(id);
+
+        if (screenwriter == null) return new ActionResponse(false, $"Screenwriter with Id {id} was not found.");
+
         try
         {
             await ScreenwritersRepository.RemoveAsync(screenwriter);
 
-            return new ActionResponse(true, "Screenwriter was removed successfully.");
+            return new ActionResponse(true, "Screenwriter was removed successful.");
         }
 
         catch (Exception exception)
         {
-            return new ActionResponse(false, $"Screenwriter was not removed. An exception occured: {exception.Message ?? exception.InnerException?.Message ?? "Undefined message"}.");
+            return new ActionResponse(false, $"Screenwriter was not removed successful. An exception occured: {exception.Message ?? exception.InnerException?.Message ?? "Undefined message"}.");
         }
     }
 }
